ref(ts): Convert more form fields to ts
This converts some more of the simple field components into ts, also changing components to functional where possible.
